fix imports for tests; fix injection tests
